requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pytest